+++ webapp/build_tools/compile.ts
@@ -1,5 +1,6 @@
 import { writeFileSync } from "fs";
 import { join } from "path";
+const package_json = require("../package.json");
 
 function syncWriteFile(filename: string, data: any) {
   writeFileSync(join(__dirname, filename), data, {
@@ -10,17 +11,46 @@ function syncWriteFile(filename: string, data: any) {
 const appKey = Bun.env.APP_CHECK_KEY;
 let toWrite: string | boolean;
 
-if((appKey!).toString() == "false") {
+if (appKey!.toString() == "false") {
   toWrite = false;
-}
-else {
+} else {
   toWrite = '"' + appKey + '"';
 }
 
+const currentdate = new Date();
+const monthLookup = [
+  "January",
+  "February",
+  "March",
+  "April",
+  "May",
+  "June",
+  "July",
+  "August",
+  "September",
+  "October",
+  "November",
+  "December",
+];
+const datetimeString =
+  monthLookup[currentdate.getMonth()] +
+  " " +
+  currentdate.getDate().toString().padStart(2, "0") +
+  ", " +
+  currentdate.getFullYear() +
+  " @ " +
+  currentdate.getHours().toString().padStart(2, "0") +
+  ":" +
+  currentdate.getMinutes().toString().padStart(2, "0") +
+  ":" +
+  currentdate.getSeconds().toString().padStart(2, "0") + " UTC-" + (currentdate.getTimezoneOffset()/60).toString();
+
 
 syncWriteFile(
-  "../src/generated_app_check_secret.ts",
-  'export const appCheckSecret: string | boolean = ' + toWrite + ';'
+  "../src/generated_app_info.ts",
+  "export const appCheckSecret: string | boolean = " +
+    toWrite +
+    `;\nexport const appVersion: string = 'v${package_json.version}';\nexport const buildDate: string = '${datetimeString}';`
 );
 await Bun.build({
   entrypoints: ["./src/index.ts"],
@@ -28,4 +58,3 @@ await Bun.build({
   minify: true, // default false
 });
 console.log("Wrote env var, built, and minified");
-