+++ webapp/src/firebase_init.ts
@@ -1,7 +1,7 @@
 import { initializeApp } from "firebase/app";
 import { ReCaptchaEnterpriseProvider, initializeAppCheck } from "firebase/app-check";
 import { initializeFirestore, persistentLocalCache, persistentMultipleTabManager } from "firebase/firestore";
-import { appCheckSecret } from "./generated_app_check_secret";
+import { appCheckSecret } from "./generated_app_info";
 import { connectFunctionsEmulator, getFunctions } from "firebase/functions";
 
 const firebaseConfig = {