+++ webapp/src/theming.ts
@@ -10,37 +10,35 @@ export const pspColors = {
   "cool-gray": "#6F727B",
 };
 
-export const datasetPlottingColors = [
-  pspColors.field,
-  pspColors["bm-gold"],
-  pspColors.aged,
-  "#FF5733",
-  "#FFC300",
-  "#FFA07A",
-  "#8CE68C",
-  "red",
-  "#ABF1BC",
-  "#FF8C00",
-  "#CFFFF6",
-  "#FFD700",
-  "#AEE7F8",
-  "#87CDF6",
-  "blue",
-  pspColors.field,
-  pspColors["bm-gold"],
-  pspColors.aged,
-  "#FF5733",
-  "#FFC300",
-  "#FFA07A",
-  "#8CE68C",
-  "red",
-  "#ABF1BC",
-  "#FF8C00",
-  "#CFFFF6",
-  "#FFD700",
-  "#AEE7F8",
-  "#87CDF6",
-  "blue",
+export function getDatasetPlottingColor(index: number) {
+  const availableLength = globalThis.plotPalletteColors.length;
+  const mod = index % availableLength;
+  return globalThis.plotPalletteColors[mod];
+}
+
+export function getPlottingColorListLength() {
+  return globalThis.plotPalletteColors.length;
+}
+
+export function getColorList() {
+  const inputFields = document.querySelectorAll(".test-field-text")! as NodeListOf<HTMLInputElement>;
+  let newColorList: string[] = [];
+  inputFields.forEach(inputField => {
+    newColorList.push(inputField.value);
+  });
+  return newColorList;
+}
+
+export function initColorList() {
+  let storageItem = localStorage.getItem("plotting_color_pallette_color_list");
+  if (storageItem == null) {
+    globalThis.plotPalletteColors = defaultPlottingColors;
+  } else {
+    globalThis.plotPalletteColors = JSON.parse(storageItem);
+  }
+}
+
+export const defaultPlottingColors = [
   pspColors.field,
   pspColors["bm-gold"],
   pspColors.aged,